Some entity fixes, some carver fixes, some other fixes

+++ src/main/java/net/tropicraft/core/common/Util.java
@@ -81,7 +81,7 @@ public class Util {
                 double d1;
                 d1 = y+0.5F - (ent.getY() + (double)ent.getEyeHeight());
 
-                double d3 = Mth.sqrt(d * d + d2 * d2);
+                double d3 = Mth.sqrt((float) (d * d + d2 * d2));
                 float f2 = (float)((Math.atan2(d2, d) * 180D) / 3.1415927410125732D) - 90F;
                 float f3 = (float)(-((Math.atan2(d1, d3) * 180D) / 3.1415927410125732D));
                 float rotationPitch = -f3;//-ent.updateRotation(rotationPitch, f3, 180D);
@@ -240,7 +240,7 @@ public class Util {
         double d0 = ent.getX() - x;
         double d1 = ent.getY() - y;
         double d2 = ent.getZ() - z;
-        return (double)Mth.sqrt(d0 * d0 + d1 * d1 + d2 * d2);
+        return Mth.sqrt((float) (d0 * d0 + d1 * d1 + d2 * d2));
     }
 
     public static Field findField(Class<?> clazz, String... fieldNames)
